The world interacts with you! :)
New:
iField-System:
The iField-System is used to interact with the player while he interacts with the world. For example if you define an iField for a stone-button and the player clicks on the stone-button, it triggers a defined command-set. In this example we can open an inventory when the player clicks on the button.